+++ docs/changelog.rst
@@ -15,6 +15,7 @@ Fixed
 ~~~~~
 
 - Resolving of relative paths in schemas. `#303`_
+- Loading string dates as ``datetime.date`` objects in YAML loader. `#305`_
 
 `0.17.0`_ - 2019-11-21
 ----------------------
@@ -450,6 +451,7 @@ Fixed
 .. _0.3.0: https://github.com/kiwicom/schemathesis/compare/v0.2.0...v0.3.0
 .. _0.2.0: https://github.com/kiwicom/schemathesis/compare/v0.1.0...v0.2.0
 
+.. _#305: https://github.com/kiwicom/schemathesis/issues/305
 .. _#303: https://github.com/kiwicom/schemathesis/issues/303
 .. _#301: https://github.com/kiwicom/schemathesis/issues/301
 .. _#295: https://github.com/kiwicom/schemathesis/issues/295